+++ packages/language-service/test/BUILD.bazel
@@ -56,6 +56,7 @@ ts_library(
         ":test_utils_lib",
         "//packages/compiler",
         "//packages/language-service",
+        "//packages/language-service:ts_utils",
         "@npm//typescript",
     ],
 )
