+++ packages/language-service/test/utils_spec.ts
@@ -9,7 +9,8 @@
 import * as ng from '@angular/compiler';
 import * as ts from 'typescript';
 
-import {getClassDeclFromDecoratorProp, getDirectiveClassLike, getPathToNodeAtPosition} from '../src/utils';
+import {getClassDeclFromDecoratorProp, getDirectiveClassLike} from '../src/ts_utils';
+import {getPathToNodeAtPosition} from '../src/utils';
 import {MockTypescriptHost} from './test_utils';
 
 describe('getDirectiveClassLike', () => {
